0027287: Enhance styles to support "Telling a Story" features
Added intertitle support for Cube Twist, Hexplode, Pump It Up and Strips.

Caption and intertitle helper functions are added to stylelib2.

For Pump It Up, we need to ensure the synchronization of segment and transition looping sequence won't be disrupted by intertitles.

For Strips, captions are enumerated at the global level (instead of segment level). This prevents multiple copies of the segment when used during strips-like transitions.