0027352: Intertitle captions are always in Corbel font
This bug was hard to see because the Constructor also sets the font to Corbel. (See EffectLing::GetCaptionsForSource "TODO" code.) To aid debugging, the font was temporary modified to Times New Roman. So now it is easier to see the bug which sets intertitle font to Corbel and segment captions font to Times New Roman. When this bug is resolved, I will remove the debugging codes.