#0027370: [arcadia] Benjy crashes because ATI X300 does not support ddx and ddy CG commands
It seems that ATI X300 has support for opengl 2.1 
That is a very bold claim. Because it implies that it has support for GLSL.
So now the cg profile selection will try to select GLSL in the profile cascade.
Let's see if GLSL has a supported equivalent for ddx and ddy.