#0027358: S00555_Essentials1
Added FILM_COLOR as a style parameter. This is where we divert from the original brushstoke. The AP6 version has a collection of sepia, desaturated, greyscale and normal color within different segments. Reveal cannot do that because of technical limitations that are cg based. So I have parametized the 4 color modes.