0026243: mES performance
changes merge from trunk rev 16548

For the case of large number of music or sync sound,
the AudioFileWithDelay object has been used to prevent
the opening of all audio stuff right at the beginning
of a save operation, since that will result in a crash.

However, the "on demand" opening mechanism for audio
wasn't closing the decoders when done with them, resulting
in the appearance of a "leak" (not an actual leak, since
the memory is indeed reclaimed when the saving is done
or cancelled). This prevents some very complex saving operations
from completing.

Surprisingly there was a mechanism put in to actually
release the resources, but it had a subtle bug in it that
prevented it from kicking in. This commit fixes that bug.

With this fix, the memory requirements during saving are,
for all practical purposes, independent of the number of
media items used in the muvee - video, photos, music,
all of them.

I'm reproducing the code comments below for completeness.