#0027358: S00555_Essentials1
I have inverted one of the gradient maps. The new layout gives a better feeling of a brushstoke.
One of the handwriting images had a piece of truncated text at the top left corner of the image. I've removed it.
The z-offsets are now in the milli- range rather than the centi- range.
The ordering of the handwritten mask has been changed.
Diffuse-glow for video files has been removed. Since we can't use Cg, I tried to fake it with Volumetric lights. But then, the alpha gets really messed up. It has been replaced with a blank effect.