0027386: MooPlayer could be useful in quickly porting aP6 styles to Reveal.
First cut at playing moo files within a Reveal style.
This commit adds a new "MooPlayer" effect which has
a single "Path" parameter which you set to the full path
of the moo file to play.

DONE: Basic support for 4:3 moos. Christmas styles
seem to work. Some photo centric moos also seem to work.

NOT DONE: Moos which use masked video
layer aren't yet supported. 16:9 mode causes
everything to be stretched. All the moo graphics
are of very low quality.

BUGS: There are quite a few rough edges such as a
scaled image not quite covering the whole screen,
wrong aspect ratio of input media, etc. The current
state is suitable for quick evaluation of whether a
particular moo is worth using in a style or not.
When we do decide to use moos in styles, the MooPlayer
can be fleshed out.