0016490: Constructor bugs - The Collectors Edition
Tweak to the constructor that provides adjustable cut hint alignment
via a text file - c:\muveeDebug\DisableCutHintValueAdjust.txt

See http://elrond/mediawiki-1.4.9/index.php/Diagnostics_in_Anchorvale

Extract -

When present (and empty), it allows the alignment of segment
boundaries to the strongest cut hint within an interval. You
can put in a floating point number indicating the extent to
which you're willing to move the segment boundary. A value
of 0.0 means ''move by as much as necessary without restraint''
and a value of 1.0 means ''moving to cause the speed change to
differ by a factor of 2 is equivalent to diminishing the strength
of a cut hint by 1.0''.