0026280: Provide option for testing "flat-bottom V-Curve"
DisableCutHintValueAdjust.txt can now specify two float parameters
separated by a space. The first one is as described in ~78315. The
second specifies a search range which defaults to 0.3 if not specified.
The search range can go from 0.0 to < 0.5.

Here is some documentation pasted in from the source code in ShotInfo.h.

float cutHintAlignmentSearchSpreadFactor; /* CHASSF */
/**<
 * Given a preliminary position for a segment boundary in the
 * to-be-constructed muvee, this factor gives the search
 * range to use for searching for a significant enough cut hint
 * to shift the boundary to.
 *
 * The factor indicates the fraction of previous and
 * next segments to include in the search. The range
 * goes from -CHASSF*prevSegDur to +CHASSF*nextSegDur.
 * Therefore the minimum value of the search range is 0.0
 * and its maximum value is 0.5 (not inclusive). Use 0.49
 * or less.
 */