0026280: Provide option for testing "flat-bottom V-Curve"
(Ported r16237 from trunk to Anderson_BR2_branch.
Note use of FileUtils2 instead of MVFileUtil.)

Looks like there's been a copy-paste error rgd the use of
the low pass filter in the multi-band-envelope class. The
filter was created and intended for use to generate cut hints,
but the high pass filter is what is used. Incidentally, the HPF
was being passed twice the number of samples it really should've
been passed.

This commit fixes the copy-paste error.


Conflicts:

	_CommonModules/SDKRuntime/MusicAnalyserFilter/music/MultiBandEnvelope.cpp