0027235: Better cutting to the beat desired in muvees
Committing the latest state of my tweaks with which I
think look better than demoed. I had to use asymmetric
biases to get this result. This still doesn't have
the "elastic adjustment" that Pete mentioned  during the
demo, which I'm still on to. (Its a bit tricky.)

I've also added a dump for SegDump-Visual.wav which
makes it easy to compare selected cut points to the cut hints
generated by the music analyzer. This file is generated when
segdump-visual.txt is present in the c:\muveeDebug folder.

I added a WriteWavFile function to AudioUtils for the above purpose.


Conflicts:

	_CommonModules/Constructor/mplib/constr/MediaPreallocator.cpp
	_CommonModules/Constructor/mplib/constr/ShotInfo.cpp
	_CommonModules/Constructor/mplib/style/GenericStyle.cpp
	_CommonModules/Constructor/mplib/style/MuseNames.h

Conflicts:

	_CommonModules/Constructor/mplib/constr/ShotInfo.cpp
	_CommonModules/Constructor/mplib/style/MuseNames.h