#26400: Features for "Telling a Story" initiative
Made optimizations to mES to use far less of the
model view matrix stack than it did before. The depth
has now reduced from 24 to 9 (vs a limit of 32).
The way this is done is to make consecutively stacked
"Translate", "Rotate" and "Scale" effects require only
a single glPushMatrix()/glPopMatrix() pair. This combining
behaviour is invisible to the style author and happens
automatically.

This fix makes it very unlikely for a style author to hit
the model view stack limit in a given style.