#26400: Features for "Telling a Story" initiative
Tweaked BACaptions.scm and Text.scm to make it easy to
replace the caption rendering scheme. BACaptions.scm now
exposes "caption-effect-maker" which is a box whose contents
you can set to a custom caption effect function for use
with layered captions as well as caption text used as
an effect.