0026400: Features for "Telling a Story" initiative
(MERGE NOTE: Didn't merge in the DFTextDemo style related changes.)

Added a "FontHeight" parameter to "Text" effect for specifying
the height of text relative to the screen height. Correspondingly
modified the S10000_DFTextDemo/data.scm to reflect the use of this
parameter.

Note that the "AutoScale" parameter now defaults to ""
instead of "x". This is in order to make the default behaviour of
"Text" as close as possible to the FTextLayout implementation.
To make it really close, PivotX must be -1.0 and PivotY must be 1.0,
but we leave those at their default values of 0.0 and 0.0
because the default rendering is centered and it is not too hard
to change that to a top-left corner pivoted rendering.


Conflicts:

	Styles/S10000_DFTextDemo/data.scm