0026400: Features for "Telling a Story" initiative
Replaced the standard text renderer for caption with the
distance field based renderer. This is still considered
"work in progress" though the text's rendering is *much*
crisper than before.

To undo the effects of this checkin and use the old TextOut
based captions renderer, replace the following line
in stylelib2.scm -
    (Captions.presenter Captions.standard)
with the following line
    (Captions.presenter ())

This line is near the end of the file.