#0026400: Features for "Telling a Story" initiative
Captions.scm
Adding CharacterFade to the list of caption animations.

stylelib2:
TubeLight is the default caption animation in stylelib2. This should never be used because if all goes according to plan, each style will overwrite this command. I've set it to TubeLight right now cos all styles are areoverwriting the caption animation to two-line-text, which was what was the default in stylelib2.