Resolved #26367: Manual magicSpot doesn't complete if transitions are present
Previous trials used a symmetric S curve - i.e. ease-in and ease-out -
for the Ken Burns effect. This created some undesirable emphasis on the
end points.

Changing the curve to an asymmetric curve where the velocity is
constant for the first half and decelerates during the second half,
creates a much more agreeable general purpose solution.

Sample rendered videos with portrait pictures illustrating what
currently happens with linear Ken Burns and with the new
"decelerating" Ken Burns can be found in T:\BugSpecificData\0026367

The "kenburns-linear-*.wmv" files are with linear Ken Burns.
The "kenburns-decel-*.wmv" files are with the decelerating Ken Burns.
You can see what slow, medium and fast paces look like for both
types of animation.

PS: Frankly, I didn't expect to resolve this bug so easily, but
am very pleased that it *can* be resolved so easily :) (-Kumar)