#0027394: Request application to block pirated keys gathered from BI data
If user entered pirated key before we release this build, and then upgrade to this build. We will check pirated key and show message, and then turn app to expiry mode.

As purchased sku missing environment variable, we need to either add them to arm file, or hard code correctly. Will need to consult YC for this....