0027412: Video highlight issue
The BugSpecificData/233267 test case captured some
problems with the preceding fixes to highlight handling
- #26036 .

1. The elimination of ultra-short shots resulted in
   some broken sequence indices in the shot vector.

2. The SplitSegment and JoinSegment operations weren't
   updating the "warped time" values based on loudness pacing.
   This wasn't necessary until these operations needed to be
   performed *after* aligning segment boundaries to cut hints.

3. The CorrectOverlappingPartitions() operation was joining the
   wrong segments in the mixed material case - an isolated video
   group in the middle of the production needed resolution by
   segment merging in order to manifest this error. This project did.

This commit resolves the core construction issues that cause
the 233267 case to fail. The issue of some highlighted segments
needing to be slowed down despite the absence of a time warp still
remains and is not as serious a problem.

This issue is mostly considered fixed by this commit, but I'll resolve
it after I get rid of the above mentioned slow downs.