##0027358: S00555_Essentials1
The handwriting fades in and out across a segment. I have modified the transparency curve (the rate at the handwriting appears and disappears) from a simple sine curve to a slightly sharper one. 

The old formula was:
transparency = sin( progress );

The new formula is:

transparency = sin( ( progress * 1.4 ) - 0.2 );

These values don't mean much. They are just sweet spots that I feel best captures the mood.

The reason I had to change the simple sine curve it because at the beginning and end of the segments, the brush-stoke transition would wipe out input A and wipe in input B. But the handwriting would still be in the screen in cases where the wipe was too fast.

There's no easy way to fix this problem. The modified sine curve will make the transition more esthetically pleasing to the eye.