RESOLVED: #0027454: [All styles] Black muvee preview with only music playing in the background #0027455: Mojave/Sunset Blvd. - Black muvee preview with only music playing in the background
#kumar took time off his holiday to do a binary search of the logs with me to find this obscure bug. tnx bro.

When we create an opengl context, we need to delete it. Intel Pineview renderers are not happy with this. So we get a black screen.
This is clearly a driver issue that muvee needs to circumvent. We basically need to hide their bugs.

Now I have modified the Cleanup() function inside GLWindowContext.cpp to check whether we are dealing with a pineview renderer; and if yes, not to delete the context.

Note that this is a hack on our side to fix their problem.

We should monitor driver updates for pineview and check if this problem persists in the future with driver updates. If it doesn't, we have to remove it.

Inside the log files, to know if we are indeed bypassing the context deletion, check out for the line:

GLWindowContext.cpp(257)           	intel Pineview detected. Bypassing wglDeleteContext...