0027439: Style misbehaviour
Fixed buggy commit r16973. On BR2, the piecewise-text-effect
function was only being used for the typewriter effect, but
on the trunk, it is being used for the two-line stuff as well,
and the empty string case wasn't accounted for, causing a crash
while trying to typeset captions containing empty strings.