Resolved #27478: Application crashes after construction muvee with Reflections style
Changing the name of the Layer class in moo/Layer.h
from "Layer" to "MooLayer" solves this crash. For
some reason, the linker linked the Layer class usage
in TextOut to moo/Layer.h despite TextOut.cpp explicitly
including pure/Layer.h. The entire implementation of Layer
is contained in the header file and therefore all includes
of Layer.g ought to be independent, but the linker doesn't
seem to think so.