0027235: Better cutting to the beat desired in muvees 0016490: Constructor bugs - The Collectors Edition
(merged trunk@17102)

Further to r17016, r17030 and r17092, this commit
limits the search range of cut hints for the first and
last segments of a muvee so that we don't get absurdly
small segments at the start and end. By the nature of the
algorithm, we bias against ultra short segments in the middle,
and can therefore have a wide search range, but the start
and end of a muvee are fixed points that can result in the
muvee beginning or ending with ultra short shots if we
don't limit the search range. The code diff is pretty simple
and commented. Please read to understand more.