#0027355: S00553_California2
OldMoviesScratches now blends smoothly with the user input. Previously there was no interaction whatsoever between the two.

The tv scanlines in California have an alpha of 0.04 for some reason. U can even seen them. I've increased it to 0.1
I've capped the scratches alpha to 0.7. It's very obvious and ugly beyond that value.
The scratches have the same color as the season.