0027355: S00553_California2
Spring colors were a bit brighter than before. Corrected Spring colors to the original calculation as of rev 16916. It was supposed to be corrected in rev 17047, but somehow i missed it.

The formula for LinearDodge should be A+B, but our formula was 2A+B-1. So we renamed our old-LinearDodge to CustomDodgeBurn. This name is chosen because of the mathematical relationship:

CustomDodgeBurn(A,B) = 2A+B-1 = LinearDodge(2A-1,B) = LinearBurn(2A,B)

Going forward, all references to the old-LinearDodge should be redirected to CustomDodgeBurn. Existing codes have been fixed since rev 15480. This commit reverts the calculation to 2A+B-1.