Resolved #27528: [intel pineview 945G] Encounter OpenGL error upon launching the application
(cherry-picked trunk c17181)

(Rgd #27448 and #27450)

Modified OpenGLInfo() implementation in mes.cpp to request the
shading language version only if the OpenGL version is >= 2.0.
In case some drivers misbehave by not providing the shading
language version even for OpenGL version >= 2.0, I discard
any errors that may be reported as a consequence of the
glGetString() call.

Test status: Haven't tested on Pineview, but pretty confident
this will fix it.