Resolved #27575: New text rendering - T/C with Typewriter text animation will be shown for 30sec before actual muvee playback
(cherry-picked trunk@17261)

Made the T&C use 100% of the time for displaying text
when using the typwriter effect. The problem resulted
from the TitleAndCredits.scm not being modified despite
the whole T&C duration being made dependent on the
amount of text to display. This commit fixes that by
making the "show-period-frac" be 1.0.