#26400: Features for "Telling a Story" initiative
The change to use "captions-for-interval" in rev 17224 caused opacity problems during transitions. It affected the majority of the styles which are stacking captions on the segment level. Reverted to use "source-captions" as originally intended.

Styles such as Strips require stacking of captions at the global level. Hence a similar method was created in Text class which uses "captions-for-interval" to get captions.

To summarize, after this change we now have two methods to stack captions:
- "source-captions-layer" for stacking at segment-effect level, and
- "captions-for-interval-interval" for stacking at global-effect level.