RESOLVED: #0027600: text rendering - credit disappear on long T/C with Typewriter animation on Kinetic Style
I take back everything I said in my previous note. This issue is not a driver issue. 

It was caused because of z-fighting.

I have moved the background effect to z = 0.005 and changed the alpha to 0.99. It fixed the issue.

At first I found it odd that the typewriter would work for titles but not credits. Cos in the scum file of Kinetic, titles and credits have the exact same code. Well it turns out that they have the exact same code on the TRUNK. not BR2. Haiz..