Resolved #27634: Highlights might get trimmed despite non-highlighted material being included
This bug is most perceptible with longish highlights -
i.e. those which are likely to span more than one segment
at the style specified cutting rate.

The step that adjusts segment boundaries to cut hints
can end up adjusting the boundary by so much that there
is no time to show a complete highlight in the segment
allocated for it. This commit fixes that by calculating
a minimum segment duration based on highlight->segment
assignment and passing it to the cut-hint based
adjustment step so that it can properly account for it.