0026243: mES performance
BUGFIX: memcpy avoidance optimization for the MCVideo
module was NOT being bypassed for interlaced video.
The problem was that the m_interlaced variable was
not set before being used. This patch addresses that by
initializing m_interlaced before taking the decision to
enable the memcpy avoidance code or not.

(This commit restores 17507 and 17511 which were
reverted in 17542. With this patch, mufat tests pass
with interlaced MC video.)