#0027602: Text rendering - truncation will appear on long T/C with Typewriter animation on Splatz Style
The text is truncated because the text layout rectangle in which the title and credits reside is too small. I have increased it to the size of the screen. To make the text appear at the top left, I'm using the "align" tag to position the text at the top right corner. However, the previous title and credits implementation had placed the text still in the top left but not at the corner as it is presently. I think it's an acceptable compromise.