Resolved #27310: Boost sountrack during selected clips and duck during VO clips
One constructor subtask was left unfinished for this issue which this
commit tries to address. The issue is that clips with the sync sound
boost setting turned on, should always be played back at normal speed
in the muvee because the audio will otherwise go out of sync.

This is a pretty hairy issue to solve completely and this commit
makes a first attempt. Quite a few cases work - including highlights
with boost applied, and boost without highlights. Corner case testing
hasn't been done yet.

Apart from corner case bugs, this feature is essentially complete. Whew!