0026243: mES performance
Vastly speeded up the PageCurl effect for the
case of lots of text. The page curling grid for
each texture is now computed based on the amount
of screen space it is occupying rather than
the fixed amount of 40x40.

With this change, a page full of text in the title
sequence of Scrapbook will perform just fine with
the new text renderer, whereas it stutters badly
with the old 40x40 grid without compromising the
visual quality.