0026243: mES performance
Further to r17606, optimized the grid even more
by using separate division count for the x and y
directions. This optimizes the grid for graphics
that are longer in one direction than the other.

The minimum number of divisions is 2, so the
smallest possible grid is 2x2. In the case of
Scrapbook with the default font size, the grid
size for each character ends up being 4x4.