Resolved #26858: Caption text, when presented, should be readable
Put in a minimum display duration for captioned photos
that depends on the length of the caption. This also applies
to inter-titles.

Currently a minimum duration of 1/2 second for every 10 characters
is assigned, with an extra 5 character count as padding. All runs
of spaces are collapsed into a single space.

The above settings look alright, but it'll help have the opinion
of a few more eyes to reduce biases.