0027257: Style tweaks
Decided to upgrade the music pacing. This is the same standard music pacing as most of our latest styles. I have adjusted the range such that it gives similar segment durations as before (using default style parameters). What is different is that the new pacing reacts better with sliders. Transition duration is also longer so that you enjoy looking at the page flips (which I think looks very nice!)

Removed code to dissolve between two short video segments.

In view of the new captions (placed on a slightly different z-offset compared to the old one), the stacking of layers have been tweaked.

Reorganized and refactored the code.