0027668: Stationary zoom-filled portrait photos is not supported
Introduced support for stationary zoom-filled portrait photos. This code is activated for portrait photos if the "Motion Effects" and "Full Portrait Pictures" settings are both set to "No". Prior to this check-in, the old behavior was to show full-portrait pictures. (This conflicted the disable "Full Portrait Pictures" settings.) The new behavior is to show a zoom-filled version of the picture as determined by the magicSpot data, but there will be no motion. The size of the rectangle used is the larger of the two magicSpot rectangles.

With this new code, all style codes that use (crop-aspect-ratio) will give inaccurate result because it did not take "Full Portrait Pictures" settings into consideration. Hence a new muSE-X function (source-content-aspect-ratio) is written to provide style authors to get the clip's actual "presentation" aspect ratio. Consequently, all shipping style codes with the old (crop-aspect-ratio) function have been modified to use the new (source-content-aspect-ratio) function.

The code to get the aspect ratio has been abstracted to stylelib2 library, along with other supporting functions. Style authors need not worry about its implementation anymore.