0026243: mES performance
Experimental "fast font caching" mechanism for text rendering
in order to reduce the preparation time for preview.

The mechanism kicks in only during preview and saving quality
remains unaffected. You can disable the "fast font cache" by
creating "c:\muveeDebug\DisableFastFontCaching.txt" to evaluate the
quality difference.

TWEAK: This commit also moves the fontcache folder from the "User data dir"
to the "common data dir" so that in the multi-user case we don't have
to regenerate characters that are already generated.

BUGFIX: A bug with GLResourceManager::RequireFont() was also fixed.
The bug was causing a doubling of the number of textures for font
characters on the average. Thankfully, the texture sizes for font
characters are insignificant compared to texture sizes for photos, so
this bug doesn't impact the VRAM usage much at all.

EXPLANATION: The "distance field maps" generated for each character
are 64x64 in size and take some time to be generated (roughly
3.4 chars/sec). By reducing this size to 32x32 during preview, the
character generation speed is increased to roughly 11 chars/sec
(> 3x improvement). However, this results in a reduced quality for
the rendered text which will be unacceptable in saved muvees.
To prevent the quality reduction, the full 64x64 characters are
rendered during saving. When the font is subsequently used
during preview, the higher quality 64x64 version is loaded.

The 32x32 sized files have the extra name fragment "_small"
attached. For instance, 'a' will be in the file "96_small.png"
and "96.png".