0026243: mES performance
Merged bugfix and tweak from trunk@17807.
Following text is copied from ~83678 -

TWEAK: This commit also moves the fontcache folder from the "User data dir"
to the "common data dir" so that in the multi-user case we don't have
to regenerate characters that are already generated.

BUGFIX: A bug with GLResourceManager::RequireFont() was also fixed.
The bug was causing a doubling of the number of textures for font
characters on the average. Thankfully, the texture sizes for font
characters are insignificant compared to texture sizes for photos, so
this bug doesn't impact the VRAM usage much at all.