0027699: Show muvee summary during Rolling Credits
Commit 17716 introduced the "rolling credits summary"
experimental feature, but introduced a bug in the
normal construction process resulting in the title
always being missed out in the visual part of the
production though the title's sound effects are present
in the production!

This commit fixes that problem, but breaks the "rolling
credits summary" experiment with the goal of keeping
the trunk in a good state all the time.

The StyleBase::MakeMuvee function has to operate in this
specific order as designed - MakeTitle->MakeCredits->MakeBody.
For the "rolling credits summary", this flow was changed to
the INCORRECT order MakeBody->MakeTitle->MakeCredits.
The reason for that original order is that when the duration
of the muvee is specified in advanced (ex: music given or
manual duration given), then the duration of the body part
that the constructor needs to target is the given duration
*LESS* the title and credits durations. Therefore the title
and credits durations need to be known *BEFORE* the body is
constructed.

Chetan - please fix your "rolling credits summary" implementation
to conform to the above constraint correctly.