0026243: mES performance
Further to r17807, I'm disabling the "fast font caching"
mechanism by default instead of enabling it by default.
So you now need to create "C:\muveeDebug\EnableFastFontCaching.txt"
to enable it.

This means that a virgin build of the trunk will have the
fast font caching disabled. I took this decision 'cos I'm
unhappy with the text quality of the 32x32 based rendering.
For example, in the default title and credits, if you look
at the character "m" in "muvee", the curves appear nearly
broken with 32x32 but are nice with 64x64.