0027257: Style tweaks
Realized that bump1.mvx is not used because the audio did not play when used as a sound effect. Could be because its duration is too short. It worked after its duration is adjusted to 1 sec. Readjusted bump2.mvx to 1 sec long as well.

bump1 gives a low-pitched "whoosh" sound, which I have replaced bump2 as the collision sound. bump2 gives more of a high-pitched "hooh" sound, which is more suitable sound for flying in and out.

Set crossfade between looping background images to 4 sec.