0026400: Features for "Telling a Story" initiative
Fixed problem with LayoutX1, where its value is 0.7 on 16:9, and 0.9 on 4:3. It happens in the muvee body only. This caused problems in captions placement and alignment.

Turned out that the dwForAspectRatio is always 0x00000000. This forced the caption aspect ratio to be always 4:3, even though the output is 16:9. Because Arcadia lets style decides on text position, we can safely set this to 0xFFFFFFFF if the font name, rectangle width and height are invalid.