#0027699: Show muvee summary during Rolling Credits
There was a bug in the show end time when we were dealing with video files. I was not properly terminating the show with a 'cut'. In the case where the credits was longer than the muvee summary, a lot of asserts were being thrown by showtime_videofile.cpp because it was seeking frames beyong the media stop time. This has been rectified.

The code has also been refactored.