0026400: Features for "Telling a Story" initiative
Implemented default-font-override feature. This allows the app to override style's default font. Originally experimented with exposing it as style parameters, but this would mean that all styles would need to include the particular style parameter. A more elegant approach is to ask the Constructor to define to a muSE symbol for the custom font, which can be overridden at various places. In this implementation, all styles just need to set its own default font. The mechanism to switch to custom font will be automatically handled by the Constructor and Captions.scm. Should the style author want to override the overridden font, it will still be possible to do so using the caption-font-override method.