RESOLVED: #0027824: Summarize to manual length doesn't work well.
If we are trying to summarize to a manual duration and that duration exceeds the total amount of video available, then the most sensible thing to do is to fall back to keep it all. That will ensure that we are as close as possible the the desired time requested by the user.