0026400: Features for "Telling a Story" initiative
Enhanced text capabilities by exposing parameters to control the following:
- anchor positions (continuous rather than discrete)
- title and subtitle scaling factors
- auto font scaling to fit layout boundaries
- shadow color, angle, distance and smoothness

Enhanced piecewise-text-effect to support line-level text scaling. This is essentially a reimplementation of first-line-scale and second-line-scale from the old two-line-text, but abstracted to be common to all caption presenters. It is also designed to be compatible with line-level horizontal alignment.

Added mvcaptioninfoex.h to mplib project. Removed duplicate stylelib2.scm.