0026400: Features for "Telling a Story" initiative
Fixed crash when typewriter animation is used in title/credits. This is because typewriter animation makes use of Captions.piecewise-text-effect, which assumes info.Shadow*, info.Scale* and info.AutoFontScale to be defined. So this fix initializes these keys to take on the same default values found in EffectLingo.

This should fix the muFAT crashes involving Turn Back Time style.