0026400: Features for "Telling a Story" initiative 0027954: [Cogen - Caption] Bringing back Reveal 8's caption feature
Further tweaked caption backward-compatibility with pre-Arcadia modules.

Added a new font info "TextAnimation" to indicate if the caption text should animate. Text animation is decided by the logfont passed in by the SDK. If the font face and height are not defined (as in the case for Arcadia two-line captions), then it means that the style will apply text animation. On the other hand, if the font face and height are passed in from the SDK, it means that the style should mimic the pre-Arcadia static text behavior.