#0027941: XMU-1540: White frame is shown on the preview with the scrap book style
This problem had not been fixed. I was doing something rather stupid in my previous 'fix'. I was not deleting the image pixels, hence keeping the texture alive. What that meant was that the image artefacts 'disapeared' because a new texture was generated for the subsequent images. But the problem used to happen when we were re-using old textures whose lifespan had been exceeded.

Anyways now what I am doing is explicitly clearing the texture once we return it to the texture pool. The issue only seems to appear on the mac so i've put the fix within an ifdef MAC64.