0027287: Enhance styles to support "Telling a Story" features
Added set-default-font-properties to allow style authors to set font color, size, face and styles conveniently in one line. Added get-font-string to allow style authors to retrieve the current LOGFONT string to be used for title-credits.

Reverted standard-title-credits z-order change done in rev 18365. It was causing z-fighting issues with Scrapbook page flip in title/credits.

Set top alignment for captions when text animation is disabled. This is to be consistent with the convention used by the app in block text positioning.