0028024: Enhance time-limited saving feature to support different locales
Modified MultiLocStringTable::LCIDToString to output the current locale string in xx-YY format. Note that the function call GetRfc1766FromLcid does not output locale in this format. So we used a dummy approach here until we find a better way.

Passed the current locale string as a parameter into logo.scm.