#0028016: [OEM - CoGen] The specification of rolling credit preview window is still in English.
In lieu of showing the muvee summary in the personalize window, what we were doing was showing a gray frame with a string in it which said:

"The muvee summary will be shown here."

That works fine in English but for other languages such as for localization, this is currently not possible because the locale of the app is not available to me when I am constructing the credits. I have spoken to Ernest and Yee Cheng and they are against sending the locale to the constructor because at teh SDK level, this will create a lot of technical issues with OEMs.

So what I have now done is to add a dummy image with film sprokets that suggests that this is a muvee frame.