0027287: Enhance styles to support "Telling a Story" features
Added StaticFade presenter, which is the same as Static but with controllable fade parameters. Although it is possible to add parameters to Static or tweak Drift values to achieve this, I have decided to create a new presenter instead.

Cleaned out remnants of TEXT_FONT which was left out from rev 18536 changes.