Resolved #0027783: Incorrect font cache folder generated when fallback font is used
Really use Arial if the required font does not exist. This prevents fontcache creation for the requested font if the font is not available. Instead, create Arial characters as fallback.

The assert check in Text.cpp is no longer required since an empty m_FontFamily string would also trigger an equivalent MVLOG in GLResourceManager::RequireFont.