#0027941: XMU-1540: White frame is shown on the preview with the scrap book style
I have halved the size of the border with no noticeable loss in quality. This will prevent the texture from being reused by a mac image. Essentially the borders will occupy 256x256 textures, which is separate from the normal images that are expected to be bigger. This will remove the artefacts that are seen in scrapbook. This is a hack.

The main problem is that user images are sharing the stock material in muvee making and the textures are not being properly cleared on the mac. I'll have to investigate why this is happening.