RESOLVED: #0028146: [OEM - CoGen] A009 – Title, Caption and credit texts can't be displayed with the style "Reflections". RESOLVED: #0027955: T/C, Intertitles & Captions are distorted in the muvee production with all styles on Amberine RESOLVED: #0027851: T/C and Inter-Titles are distorted in the muvee production when Reflection style is selected RESOLVED: #0028102: T/C, Intertitles & Captions are distorted in the muvee production with all styles on Celeron 2.66
I have increased the opengl-shading-language requirement from 1.2 to 1.3.
This is because most (if not all) on-board intel graphics cards claim to support opengl 2.0 when in fact the drivers are not properly configured for that.
If the user has a separate graphics card, it is fair to say that they will support opengl-shading-language 1.3 and will thus be able to run the high-resolution text.