Resolved #28144 key cause is that we're no longer able to release the threading class immediately after analysis after porting to use the callback mechanism. the check for the availability of the threading class pointer to assume "analysing in progress" is thus no longer valid.
hence when the same media is made to analyse a 2nd time during a session (only pictures do that after rotation)... the analysis doesn't start for the 2nd invocation.

Have fixed this by checking the class instance for the availability of the analysis module instead since the module is gauranteed to be released when analysis is done.