0026400: Features for "Telling a Story" initiative
Reverted rev 17626 changes because that caused inconsistency with the app's base height. It is evident if you compare the font preview in the caption dialog with the make-muvee output. The make-muvee output font size is 33% smaller. After undoing this change, the font heights (relative to the output frame height) between title, credits and the caption dialog are all standardized.

Fixed a bug in Typewriter effect which causes cursor color to be always white.

Fixed an issue with Rolling text animation process where it was not monotonically increasing when the font height is greater than the layout height.

Fixed another issue with Rolling text fallback effect using TextOut, where the text did not roll completely out of the frame in Classic Vanilla and Ultra Plain styles.