RESOLVED: #0025826: Preview muvee shows black screen after applying customised background to T/C
The Intel Pineview renderer crashes if we try to delete an opengl context and then create a brand new one. It's specifically unhappy about the

wglDeleteContext(..) function call.

Here's the problem:

When we click on the Personalize window a preview a text effect of background effect, We create and delete an opengl context, which is perfectly fine and a very sensible thing to do if you don't want any memory leaks. Intel Pineview thinks otherwise and doesn't allow new openGL contextes to be created once we deleted the previous one. 

Ok this is a driver issue. We need to monitor driver updates and then hopefully in the future, we can remove this check.