Resolved #0028257: [OEM - CoGen] U002 – Some font change can't be applied to caption, intertitle, title, credits and logo (footnote message)
Forced the LOGFONT string construction to use default CharSet, OutPrecision, ClipPrecision, Quality and PitchAndFamily fields, i.e. "1,0,0,0,0". This is consistent with the LOGFONT string as returned by the font dialog box. This explains why certain fonts (e.g. ABOUTFACE) can be rendered correctly in the font preview window, but not in the muvee output.