0028341: Crash in Turn Back Time title/credits preview with the old text renderer upon first launch
Moved Glew out of the block below so that the
Glew context is also available during the
ReleaseResourcePools() call at the end. There
is a stray case in which ReleaseResourcePools()
ends up allocating a texture when previewing
Turn Back Time style's title/credits as soon as
you launch muvee Reveal. Making Glew available
to ReleaseResourcePools() solves that problem,
though we need to look into why such a texture
is being created there.

The problem happens only when forcing the use of
the old (non-shader) text renderer.