0028384: Saving has high memory/cpu usage with rolling credits on single core machine
Change merge from trunk rev 19343

BUGFIX: Fixed the uncontrolled growth in number of video server instances when
saving a muvee featuring lots of video clips and rolling summary at the end.
(The problem happens only on single core machines.)

The fix is to mark the video file usage in the rolling summary with a special
identifier "*rolling-summary" (which will be ignored by MES) so that those
video files are treated as "different" and therefore assigned by VideoPool
to different video file pools. Had the VideoPool allocated them to the
same pool, the video file objects will need to be kept alive for the entire
duration of the muvee on single core machines.

On multi-core machines, this isn't a problem, since the VideoFileWithDelay
mechanism is used, which only keeps the needed number of video files
alive. This is possible because files can be opened and closed with
comparatively little impact on preview or save performance in the case
of multi-core machines, but on single-core machines such killing and
re-opening of video files can kill performance.