#0028246: A tool is needed to judge the performance of individual machines to figure out what the best performance will be.
Added a verbose flag to display more info.

The output is now saved into a text file with the name of the driver version in the same folder as the executable. The driver version was chosen because it is very unique and will change if the OS, graphics card and driver changes. 

Minor code formatting