#0028246: A tool is needed to judge the performance of individual machines to figure out what the best performance will be
While saving, we force mipmapped to be on.
The threshold for choosing a better candidate has been changed to 10% from 5%.
The first result in the profile selection cascade will be mipmapped.

A post build step has been added to the project file so that the executable and test image will be copied to the output folder.