RESOLVED: #0028589: [OEM - CoGen] On old text rendering system, using zoom in zoom out animation, T/C will zoom in zomm out and then zoom in again.
The minimum character height has been set to -2 in the MuseBasedTextStyle. The main issue here is integer interpolation. In the case of zoom-in and zoom-out, we animate from a maximum height of -20 to a minimum height of -1. But because we are dealing with integer division, if the muvee progress exceeds 1.0 by a minute amount ( which happens quite often), the character height is rounded down to 0, which is an undefined character height.

This issue has been present for years. We never noticed it because it happens at the very end of the title animation and is usually masked by the begining of the muvee which starts to fade in.