0026400: Features for "Telling a Story" initiative
Added ability to override OpenGL Shading Language version number check though a muveeDebug file: OverrideShadingLanguageVersionRequired.txt.

You can specify the minimum version number for which your graphics card need to meet in order to use the new text rendering. If this file does not exist, or if the version number cannot be extracted, then it will use the default value of 1.3.

You can therefore make use of this to switch between old and new text renderer. If you put 0 as its content, this will always pass the ShadingLanguageVersion check and therefore use the new text renderer. On the other hand, if you put a high number such as 9, this will always fail the version check and fallback to the old text renderer.