0027939: Application crashes while saving muvee to MOV-HD 1920x1080
changes merge from trunk rev 19346


Saving to MOV format have the Window of 10 frames before it output the compressed frames into the file
So  QuickTime component keep 81 MB of data into the memory before output the frame.
This thing cause component to run out of memory on some machine.

Reducing the Windows size to 4 will help to avoid this situation for HD output.