Resolved #0028684: The media path of BumperPix and Techno Buzz is missing a backslach on the trunk
With the previous change, the muSE function (sound-fx ...) needs to have a preceding backslash in the path, while (audio-clip ...) need not have it. This leads to a lot of inconsistencies for style authors in specifying sound effects.

Other styles that uses sound effects (e.g. Hexplode, Strips, Turn Back Time, etc.) are unaffected by this issue. Only Grunge and Techno Buzz are affected. Only these two styles use (sound-fx ...) muSE function.

The root cause is at SoundFX.cpp, where the base path is assumed to end with a backslash. This is no longer an assumption after the Constructor backporting effort. We need to use MVPathUtil::StdPathAppend instead.