#0028246: A tool is needed to judge the performance of individual machines to figure out what the best performance will be
This is kind of a major change because of the number of places where the code has changed. But in a nutshell, we no longer use strings and solely use wstrings. However the encoding is in UTF-8. I dunno if that'll to for localized paths. I'll do a test to find out.