#0028246: A tool is needed to judge the performance of individual machines to figure out what the best performance will be
Added using namespace std; which allowed me to delete all the std:: prepends.

mvPerformance will now accept a string argument at the command line. This string should be a fully qualified path with no trailing back-slash and will be used as the path where the xml file will be stored.

If such an argument is not present, the xml file will be saved in the same directorry as the executable.

BUGFIX: When the algorithm was finding the absolute difference between the current readback/upload time and the desired time, abs was used instead of fabs. This has been fixed. Btw this bug is quite a major one and has compromised all the data i've been collecting so far. I might have to re-run the app on all machines and relook at the data.